+++ lib/ui/profile/ProfileInfoPage.dart
@@ -31,11 +31,10 @@ class ProfileInfoPage extends StatefulWidget {
 
   final ProfileInfo contentType;
   final Map<String, dynamic>? params;
-  final bool showProfileCommands;
-  final bool showAccountCommands;
+  final bool onboarding;
   final void Function()? onStateChanged;
 
-  ProfileInfoPage({super.key, required this.contentType, this.params, this.showProfileCommands = true, this.showAccountCommands = false, this.onStateChanged});
+  ProfileInfoPage({super.key, required this.contentType, this.params, this.onboarding = false, this.onStateChanged});
 
   @override
   State<StatefulWidget> createState() => ProfileInfoPageState();
@@ -62,18 +61,11 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
   bool _updatingDirectoryVisibility = false;
   bool _preparingDeleteAccount = false;
 
-  bool get isEditing => _editing;
+  bool get _showProfileCommands => (widget.onboarding == false);
+  bool get _showAccountCommands => (widget.onboarding == false);
+
   bool get isLoading => _loading;
   bool get directoryVisibility => (_privacy?.public == true);
-
-  void setEditing(bool value) {
-    if (mounted && (_editing != value)) {
-      setState(() {
-        _editing = value;
-        widget.onStateChanged?.call();
-      });
-    }
-  }
 
   Future<bool> saveEdit() =>
     _profileInfoEditKey.currentState?.saveEdit() ?? Future.value(false);
@@ -115,14 +107,17 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
 
         if (directoryVisibility == true)
           Column(children: [
-            Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
-              Text(_desriptionText, style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
-            ),
+            if (widget.onboarding == false)
+              Padding(padding: EdgeInsets.only(top: 16), child:
+                Text(_desriptionText, style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
+              ),
 
-            _editing ? _editContent : _previewContent,
+            Padding(padding: EdgeInsets.only(top: 16), child:
+              _editing ? _editContent : _previewContent,
+            )
           ]),
 
-        if (widget.showAccountCommands && !_editing)
+        if (_showAccountCommands && !_editing)
           _accountCommands,
       ],);
     }
@@ -136,11 +131,12 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
           contentType: widget.contentType,
           profile: _profile,
           privacy: _privacy,
+          onboarding: widget.onboarding,
           pronunciationAudioData: _pronunciationAudioData,
           photoImageData: _photoImageData,
           photoImageToken: _photoImageToken,
         ),
-        if (widget.showProfileCommands)
+        if (_showProfileCommands)
           Padding(padding: EdgeInsets.only(top: 24), child:
             _previewCommandBar,
           ),
@@ -153,10 +149,10 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
       contentType: widget.contentType,
       profile: _profile,
       privacy: _privacy,
+      onboarding: widget.onboarding,
       pronunciationAudioData: _pronunciationAudioData,
       photoImageData: _photoImageData,
       photoImageToken: _photoImageToken,
-      showProfileCommands: widget.showProfileCommands,
       onFinishEdit: _onFinishEditInfo,
   );
 
@@ -204,9 +200,13 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
       Localization().getStringEx('panel.profile.info.directory_visibility.public.text', 'Public') :
       Localization().getStringEx('panel.profile.info.directory_visibility.private.text', 'Private');
 
-    final String messageTemplate = directoryVisibility ?
-      Localization().getStringEx('panel.profile.info.directory_visibility.public.description', 'Your directory visibility is set to $visibilityMacro. Anyone on or off the User Directory can view your account.') :
-      Localization().getStringEx('panel.profile.info.directory_visibility.private.description', 'Your directory visibility is set to $visibilityMacro. Your account is available only to you.');
+    final String messageTemplate = widget.onboarding ?
+      (directoryVisibility ?
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.onboarding.public.description', 'Your directory visibility is set to $visibilityMacro. The information below will be visible to the users of the ${AppTextUtils.appTitleMacro} App.') :
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.onboarding.private.description', 'Your directory visibility is set to $visibilityMacro. Your profile is visible only to you.')) :
+      (directoryVisibility ?
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.normal.public.description', 'Your directory visibility is set to $visibilityMacro. Anyone on or off the User Directory can view your account.') :
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.normal.private.description', 'Your directory visibility is set to $visibilityMacro. Your account is available only to you.'));
 
     final List<String> messages = messageTemplate.split(visibilityMacro);
     List<InlineSpan> spanList = <InlineSpan>[];
@@ -509,7 +509,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
         _pronunciationAudioData = pronunciationAudioData;
       }
 
-      if (widget.showProfileCommands) {
+      if (_showProfileCommands) {
         _editing = false;
       }
       widget.onStateChanged?.call();